requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
requests==2.31.0
python-dotenv==1.0.0
jinja2==3.1.3
